+++ outputs.tf
@@ -54,6 +54,11 @@ output "forseti-server-storage-bucket" {
   value       = "${module.server.forseti-server-storage-bucket}"
 }
 
+output "forseti-cloudsql-connection-name" {
+  description = "Forseti CloudSQL Connection String"
+  value = "${module.server.forseti-cloudsql-connection-name}"
+}
+
 output "suffix" {
   description = "The random suffix appended to Forseti resources"
   value       = "${local.random_hash}"