+++ CHANGELOG.md
@@ -6,7 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Extending the adopted spec, each change should have a link to its corresponding pull request appended.
 
-## [Unreleased]
+## [v2.2.0] - 2019-06-20
+
+### Added
+
+- Support for Forseti v2.17.0 [#184]
+- Add Kubernetes resources to CAI asset inventory
+
+### Changed
+
+- Make Forseti Real-Time Enforcer VM have private IP by default. [#180]
+- Do not add email configuration if `sendgrid_api_key` is unset [#174]
 
 ### Changed
 
@@ -184,6 +194,8 @@ Extending the adopted spec, each change should have a link to its corresponding 
 [v2.0.0]: https://github.com/terraform-google-modules/terraform-google-forseti/compare/v1.6.0...v2.0.0
 [v2.1.0]: https://github.com/terraform-google-modules/terraform-google-forseti/compare/v2.0.0...v2.1.0
 
+[#180]: https://github.com/forseti-security/terraform-google-forseti/pull/180
+[#174]: https://github.com/forseti-security/terraform-google-forseti/pull/174
 [#170]: https://github.com/forseti-security/terraform-google-forseti/pull/170
 [#163]: https://github.com/forseti-security/terraform-google-forseti/pull/163
 [#157]: https://github.com/forseti-security/terraform-google-forseti/pull/157